Fix file paths of actual assets and resize for yes-no game tutorial
Moved files into the correct folders to reflect the file paths indicated in the tutorials so that the game may be run from within the repository.

The asset files have also been resized to actually conform to the "length of 400 pixels for horizontal images" stated in Step 2. Else, the images would be too big and overlap. Original pictures (at size 800 pixels horizontally) have been to a new folder named "Originalimages". These may be removed if deemed unnecessary