Change character initial .y position to 620 from 310 in glod-game tutorial
The first instance of .y, from step 4, states that we want to move Glod's position from the center of the screen, and bring him down to Earth but changing his position.

Changing 'Anchor.Center' to 'Anchor.Top' then moving glod to the bottom of the screen requires `620.0 >= .y` instead of `310.0 >= .y`. This value of 310.0 changes to 620.0 without indication suddenly in hte middle of Step 4.

We simply change all instances of .y to have the value 620.0 from the start.